Move parser test to data module
